requirements.txt
flask
torch 
torchvision 
pillow